Rename run to cli and write default config.
