+++ src/modelplane/__init__.py
@@ -1,0 +1,3 @@
+from modelgauge.config import write_default_config
+
+write_default_config()
